JobsTableViewController SwipeActions Extension
